% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{add.ASC}
\alias{add.ASC}
\title{Calculate ASC score}
\usage{
add.ASC(results, column, ...)
}
\arguments{
\item{results}{table of pairwise interactions}
}
\value{
a dataframe with ASC scores
}
\description{
Calculate ASC score
}
