% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_table.R
\name{add.patterns}
\alias{add.patterns}
\title{Add some other statistics}
\usage{
add.patterns(results, patterns, name)
}
\arguments{
\item{results}{table of pairwise interactions}

\item{r.als}{random AL statistics}
}
\value{
a dataframe with additional info
}
\description{
Add some other statistics
}
