% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_table.R
\name{add.r.info}
\alias{add.r.info}
\title{Add some other statistics}
\usage{
add.r.info(results, r.als)
}
\arguments{
\item{results}{table of pairwise interactions}

\item{r.als}{random AL statistics}
}
\value{
a dataframe with additional info
}
\description{
Add some other statistics
}
