% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{AIC.SELECT}
\alias{AIC.SELECT}
\title{Calculate AIC}
\usage{
\method{AIC}{SELECT}(obj, type = "SELECT")
}
\arguments{
\item{obj}{Fitted SELECT model}

\item{type}{Use type="poisson" to get Poisson AIC. This is calculated from the
matrices of observed and fitted counts and does not condition on row totals.}
}
\description{
Calculates AIC of SELECT object. By default it provides the
binomial SELECT AIC. For comparison with models fitted by glm of gam use
type="poisson" to get the Poisson AIC.
}
