% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{al.pairwise.alteration.custom_measure}
\alias{al.pairwise.alteration.custom_measure}
\title{Generic interface to calculate pairwise scores between all the pairs of alterations. Used as front-end for MI measures}
\usage{
al.pairwise.alteration.custom_measure(al, als, f = corrected.binary.MI, ...)
}
\arguments{
\item{al}{The alteration landscape}

\item{alp}{Observed AL Overlap statistics}
}
\value{
Pairwise ME P.vlaue
}
\description{
Generic interface to calculate pairwise scores between all the pairs of alterations. Used as front-end for MI measures
}
