% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{al.pairwise.alteration.ME_p}
\alias{al.pairwise.alteration.ME_p}
\title{Compute the ME Pvalue starting from overlap statistics}
\usage{
al.pairwise.alteration.ME_p(alp, r.alp, enhanced.version = TRUE)
}
\arguments{
\item{alp}{Observed Overlap statistics}

\item{r.alp}{Random Overlap statistics}
}
\value{
Pairwise ME P.vlaue
}
\description{
Compute the ME Pvalue starting from overlap statistics
}
