% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{al.pairwise.alteration.stats}
\alias{al.pairwise.alteration.stats}
\title{Compute overlap statistics for the AL}
\usage{
al.pairwise.alteration.stats(al, als = NULL, do.blocks = FALSE)
}
\arguments{
\item{al}{Alteration landscape}
}
\value{
Pairwise statistics
}
\description{
Compute overlap statistics for the AL
}
