% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{al.stats}
\alias{al.stats}
\title{Compute simple statistics for the AL, including accurate statistics on each sample/alteration block}
\usage{
al.stats(al)
}
\arguments{
\item{al}{Alteration landscape}
}
\value{
An Alteration Landscape Statistics (ALS) object
}
\description{
Compute simple statistics for the AL, including accurate statistics on each sample/alteration block
}
