% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{am.pairwise.alteration.coverage}
\alias{am.pairwise.alteration.coverage}
\title{Compute pairwise coverage}
\usage{
am.pairwise.alteration.coverage(overlap_M, M.stats)
}
\arguments{
\item{am}{Alteration Matrix}
}
\value{
The pairwise overlap between the alterations in 'dist' format
}
\description{
Compute pairwise coverage
}
