% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{am.pairwise.alteration.overlap}
\alias{am.pairwise.alteration.overlap}
\title{Compute pairwise overlap}
\usage{
am.pairwise.alteration.overlap(am)
}
\arguments{
\item{am}{Alteration Matrix}
}
\value{
The pairwise overlap between the alterations in 'dist' format
}
\description{
Compute pairwise overlap
}
