% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{am.stats}
\alias{am.stats}
\title{Calculate simple statistics for an am}
\usage{
am.stats(am)
}
\arguments{
\item{am}{Alteration matrix}
}
\value{
Statistics (AMS object)
}
\description{
Calculate simple statistics for an am
}
