% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{APC.block}
\alias{APC.block}
\title{APC correction. Extended version, aware of subdivision of the dataset.}
\usage{
APC.block(
  results,
  divide.by = c("int_type", "direction"),
  column = "score",
  out.column = "APC",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{results}{table of pairwise interactions}

\item{column}{Which column to use as score}

\item{do.ASC}{whether calculating APC or ASC}

\item{do.fix_positivity}{Whether setting negative scores to 0}

\item{use.median}{Use median instead of average}
}
\value{
APC scores
}
\description{
APC correction. Extended version, aware of subdivision of the dataset.
}
