% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{APC}
\alias{APC}
\title{APC correction. Core version, unaware of any subdivision of the dataset.}
\usage{
APC(
  results,
  do.ASC = FALSE,
  do.fix_positivity = FALSE,
  do.asinh.before = FALSE,
  do.asinh.after = FALSE,
  use.median = TRUE,
  column = "score",
  out.column = "APC",
  verbose = FALSE
)
}
\arguments{
\item{results}{table of pairwise interactions}

\item{do.ASC}{whether calculating APC or ASC}

\item{do.fix_positivity}{Whether setting negative scores to 0}

\item{use.median}{Use median instead of average}

\item{column}{Which column to use as score}
}
\value{
APC scores
}
\description{
APC correction. Core version, unaware of any subdivision of the dataset.
}
