% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{BIC.SELECT}
\alias{BIC.SELECT}
\title{Calculate BIC}
\usage{
\method{BIC}{SELECT}(obj, type = "SELECT")
}
\arguments{
\item{obj}{Fitted SELECT model}

\item{type}{Use type="poisson" to get Poisson BIC. This is calculated from the
matrices of observed and fitted counts and does not condition on row totals.}
}
\description{
Calculates BIC of SELECT object. By default it provides the
binomial SELECT BIC. For comparison with models fitted by glm of gam use
type="poisson" to get the Poisson BIC.
}
