% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{binary.MI}
\alias{binary.MI}
\title{Binary Mutual Information. Works on single units, vectors or matrices}
\usage{
binary.MI(n_00, n_01, n_10, n_11, n_f1, n_f2, tolerance = 1e-05)
}
\value{
MI
}
\description{
Binary Mutual Information. Works on single units, vectors or matrices
}
