% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{binary.MI_debug}
\alias{binary.MI_debug}
\title{Debug version of Binary Mutual Information. Works on single units, vectors or matrices.}
\usage{
binary.MI_debug(n_00, n_01, n_10, n_11, n_f1, n_f2)
}
\value{
debug structure containing MI and debug variables
}
\description{
Debug version of Binary Mutual Information. Works on single units, vectors or matrices.
}
