% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_random.R
\name{birewireBlocEvents}
\alias{birewireBlocEvents}
\title{Generate random binary matrix}
\usage{
birewireBlocEvents(
  M,
  Q = 100,
  row.blocks = NULL,
  col.blocks = NULL,
  verbose = FALSE,
  switch_threshold = 15,
  seed = NULL
)
}
\arguments{
\item{am}{Input binary matrix (no row/col names!)}

\item{params}{parameter list}
}
\value{
One random binary matrix
}
\description{
Generate random binary matrix
}
