% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{boot.SELECT}
\alias{boot.SELECT}
\title{Implement bootstrap}
\usage{
boot.SELECT(
  Data,
  statistic,
  N,
  SetID = "Haul",
  BlockID = NULL,
  GearID = NULL,
  Freqs = c("nfine", "nwide"),
  within.resamp = T,
  ...
)
}
\description{
Apply the double bootstrap and return a statistic from each
bootstrap sample in a matrix.
}
