% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_table.R
\name{build.interaction.table}
\alias{build.interaction.table}
\title{Build interaction table}
\usage{
build.interaction.table(al, als, r.alp, patterns = NULL)
}
\arguments{
\item{al}{Alteration Landscape}

\item{als}{AL statistics}

\item{r.alp}{random AL statistics}

\item{patterns}{patterns}
}
\value{
a dataframe with all interaction info pooled together
}
\description{
Build interaction table
}
