% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlFuncs.R
\name{ccfit}
\alias{ccfit}
\title{Fit logistic or Richards seln curve to covered-codend data.}
\usage{
ccfit(
  catch = catchdat,
  type = "logit",
  probs = c(0.25, 0.5, 0.75),
  x0 = c(-10, 0.3),
  delta = 1,
  plots = T,
  suff.big = 3,
  error.bars = F,
  plotlens = NULL,
  details = F,
  main = c("Proportion of catch in large mesh codend", "Deviance residuals"),
  xlab = "Length (cm)",
  ylab = "Propn retained"
)
}
\arguments{
\item{catch}{Matrix with 3 columns. These must be
\enumerate{
\item length (i.e., midpoint of length class);
\item Freq in experimental gear; 3) Freq in nonselective cover
}}

\item{type}{Character. Use type="rich" to fit Richards curve}

\item{probs}{Numeric to specify retention probabilities for which corresponding
lengths are required}

\item{plotlens}{Numeric. Lengths are which to predict retention probabilities.}

\item{details}{If TRUE, returns additional output}
}
\value{
converged Convergence code from nlm minimizer

x Selection curve parameters, a, b, and delta (if Richards curve)

l Log-likelihoods of fitted, null and full models

lens Lengths of 25\%, 50\% and 75\% retention (by default)

sr Selection range

CF Over-dispersion correction factors, using both deviance and Pearson chisq.
}
\description{
Fit logistic or Richards seln curve to covered-codend data.
}
