% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casey_covariates.R
\name{combine_df}
\alias{combine_df}
\title{Create single dataframe of raw data, focal statistics and shapefile}
\usage{
combine_df(cent_lyr, output_slope, focal_3, focal_5, focal_7, focal_9)
}
\arguments{
\item{cent_lyr}{the SpatialPointsDataFrame from the spatial_join function}

\item{output_slope}{dataframe of calculated slope from the raster_slope function}

\item{focal_3}{dataframe of focal statistics for a window width of 3 from the focal_stats function}

\item{focal_5}{dataframe of focal statistics for a window width of 5 from the focal_stats function}

\item{focal_7}{dataframe of focal statistics for a window width of 7 from the focal_stats function}

\item{focal_9}{dataframe of focal statistics for a window width of 9 from the focal_stats function}
}
\value{
focal_data
}
\description{
Bind the SpatialPointsDataFrame, slope and focal statistics for all window widths
into a single dataframe and format column names
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
