% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{corrected.binary.MI}
\alias{corrected.binary.MI}
\title{Weighted MI measure}
\usage{
corrected.binary.MI(
  als,
  cache = NULL,
  f_correction_algorithm = f_sigmoid_max_block_correction,
  ...
)
}
\arguments{
\item{al}{The alteration landscape}

\item{alp}{Observed AL Overlap statistics}
}
\value{
Pairwise weighted MI
}
\description{
Weighted MI measure
}
