% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{dBinom}
\alias{dBinom}
\title{Extended binomial density function that also works for non-integer counts}
\usage{
dBinom(y, n, p, log = FALSE)
}
\description{
Extended binomial density function that also works for non-integer counts
}
