% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELECT.R
\name{deltamethod}
\alias{deltamethod}
\title{delta method function}
\usage{
deltamethod(g, mean, cov, ses = TRUE)
}
\description{
Returns variance matrix of functions of estimated values
e.g., for std errors of L50 and SR from a and b parameterization
}
