% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC_threshold.R
\name{establish_APC_threshold}
\alias{establish_APC_threshold}
\title{Calculate ASC score}
\usage{
establish_APC_threshold(alpi)
}
\arguments{
\item{alpi}{table of pairwise interactions}
}
\value{
a dataframe with stats on thresholds and the selected threshold
}
\description{
Calculate ASC score
}
