% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_FDR.R
\name{estimate_FDR}
\alias{estimate_FDR}
\title{Core FDR estimation}
\usage{
estimate_FDR(real_scores, random_scores, invert.scores = FALSE, max.score = NA)
}
\arguments{
\item{temp}{list of m matrices (n x n)}
}
\value{
n lists of n vectors of m elements
}
\description{
Core FDR estimation
}
