% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_FDR.R
\name{estimate_random_Pval}
\alias{estimate_random_Pval}
\title{Get random P.values}
\usage{
estimate_random_Pval(r_MI, invert.scores = TRUE)
}
\arguments{
\item{temp}{list of m matrices (n x n)}
}
\value{
n lists of n vectors of m elements
}
\description{
Get random P.values
}
