% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{Estimates}
\alias{Estimates}
\title{Returns selectivity parameters}
\usage{
Estimates(fit, OD = NULL)
}
\description{
Calculates the estimates and standard errors for
parametric selectivity curves fitted by SELECT.
The user can edit this function to add new
calculations for any user-supplied retention curves.
}
