% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Expt4Tailor}
\alias{Expt4Tailor}
\title{Bycatch reduction data for tailor}
\format{
dataframe \code{Expt4Tailor} with 730 rows and 8 columns.
}
\usage{
data(Expt4Tailor)
}
\description{
Tailor catch data (including sampling fractions) from double-rigged trawls with
panaeid gears fitted with either Nordmore grid or blubber chute.
}
\examples{
data(Expt4Tailor)
}
\references{
Broadhurst & Millar (2018). Relative performance of two bycatch
reduction devices for excluding juvenile mulloway, Argyrosomus japonicus and
other fish from south-eastern Australian penaeid trawls.
}
\keyword{datasets}
