% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{f_corr_fix_N_consistency}
\alias{f_corr_fix_N_consistency}
\title{Consistency enforcer for correction function for weight MI measure}
\usage{
f_corr_fix_N_consistency(N, marginal_f1, marginal_f2, oldN, pseudo_count = 1)
}
\value{
Corrected N
}
\description{
Consistency enforcer for correction function for weight MI measure
}
