% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{f_sigmoid_max_block_correction}
\alias{f_sigmoid_max_block_correction}
\title{Block aware correction function for weight MI measure}
\usage{
f_sigmoid_max_block_correction(
  marginal_f1,
  marginal_f2,
  real.N,
  rho = 0.1,
  lambda = 15,
  block_als
)
}
\value{
Corrected N
}
\description{
Block aware correction function for weight MI measure
}
