% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{f_sigmoid_max_correction}
\alias{f_sigmoid_max_correction}
\title{Correction function for weight MI measure}
\usage{
f_sigmoid_max_correction(
  marginal_f1,
  marginal_f2,
  real.N,
  rho = 0.1,
  lambda = 15
)
}
\arguments{
\item{marginal_f1}{Marginal frequency of event 1}

\item{marginal_f2}{Marginal frequency of event 2}

\item{real.N}{Observed sample size}
}
\value{
Corrected N
}
\description{
Correction function for weight MI measure
}
