% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_FDR.R
\name{FDR_analysis}
\alias{FDR_analysis}
\title{Perform FDR analysis and determine optimal Pvalue threshold}
\usage{
FDR_analysis(alpi, r.MI, column = "wMI_p.value", FDR.thresh = 0.1)
}
\value{
FDR estimates, optimal thresholds, annotated alpi table
}
\description{
Perform FDR analysis and determine optimal Pvalue threshold
}
