% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_FDR.R
\name{FDR_threshold}
\alias{FDR_threshold}
\title{Determine optimal P value thresholds to have a FDR.thresh (default: 0.1) FDR}
\usage{
FDR_threshold(FDR, FDR.thresh = 0.1)
}
\arguments{
\item{temp}{list of m matrices (n x n)}
}
\value{
n lists of n vectors of m elements
}
\description{
Determine optimal P value thresholds to have a FDR.thresh (default: 0.1) FDR
}
