% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL.R
\name{filter.al}
\alias{filter.al}
\title{Filter AL}
\usage{
filter.al(al, params = NULL)
}
\arguments{
\item{al}{The alteration landscape}

\item{params}{parameter list}
}
\value{
An Alteration Landscape (AL) object with the filtered am
}
\description{
Filter AL
}
