% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeSelnFuncs.R
\name{FitGAM}
\alias{FitGAM}
\title{Fit smooth function to catch share data}
\usage{
FitGAM(
  Catch,
  quasi = F,
  bs = "cr",
  k = 5,
  m = NA,
  sp = NULL,
  null = F,
  rm.zeros = T
)
}
\arguments{
\item{Catch}{Matrix with data in first three columns}

\item{quasi}{Logical, whether to apply quasibinomial correction}

\item{bs}{Choice of smoother}

\item{k}{Dimension of the basis. k=3 in minimum for natural cubic spline.}

\item{m}{Order of the penalty}

\item{sp}{Supplied smoothing parameter}
}
\value{
List containing fitted model.
}
\description{
Fit spline to catch share
}
