% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeSelnFuncs.R
\name{fitPH}
\alias{fitPH}
\title{Fit logistic seln curve to paired-haul data}
\usage{
fitPH(RawDf, fitTots = T)
}
\arguments{
\item{fitTots}{Logical indicating whether to fit to totals (over hauls)}

\item{Catch}{Matrix including columns "freqB", "SF.B", "freqG" and "SF.G"}
}
\description{
Fit logistic seln curve to paired-haul data to raw subsampled data
}
