% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeSelnFuncs.R
\name{fitPoly}
\alias{fitPoly}
\title{Fit model averaged 4th order polynomial to raw subsampled data}
\usage{
fitPoly(RawDf, fitTots = T, ...)
}
\arguments{
\item{fitTots}{Logical indicating whether to fit to totals (over hauls)}

\item{Catch}{Matrix including columns "freqC", "SFC", "freqT" and "SFT"}
}
\description{
Fit model averaged 4th order polynomial to raw subsampled data
}
