% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casey_covariates.R
\name{focal_stats}
\alias{focal_stats}
\title{Calculate focal statistics for raw SELECT model data}
\usage{
focal_stats(attribute_tble, variable_arr, size)
}
\arguments{
\item{attribute_tble}{dataframe of raw data from SELECT model}

\item{variable_arr}{vector of variables listed in attribute table}

\item{size}{moving window size for calculating focal statistics}

\item{output_slope}{dataframe of calculated slope from the raster_slope function}
}
\value{
focal_data
}
\description{
Calculate mean, std, std position and slope range from SELECT
model attribute table for a single window width (parameter: "size")
}
\details{
Function must be calculated for all window widths (3,5,7,9) to run combine_df
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
