% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/script_SpatialModel_projection_2020&on_package.R
\name{gam_dataframe_2020}
\alias{gam_dataframe_2020}
\title{Create dataframe from list of GAM models for 2020}
\usage{
gam_dataframe_2020(
  ISO3ExceptList = "ISO3s_exceptions.csv",
  selectedFeatures = "temp_SelectedFeatures_projection.RData",
  ISO3List = "ISO3s.csv",
  estDF,
  TPppBUList = "data_2000TPppBU.csv"
)
}
\arguments{
\item{ISO3ExceptList}{dataframe; list of ISO3 exceptions}

\item{selectedFeatures}{dataframe; created by update_pca()}

\item{ISO3List}{dataframe; list of ISO3}

\item{estDF}{estDF dataframe; created by general_trends() and expanded by update_pca()}

\item{TPppBUListdataframe;}{TODO}
}
\value{
list; extracted components from update Grids data => list(estDF, alloDF)
}
\description{
Extract values from GAM model list by country and calculate new fields
to create new estDF and alloDF dataframes
}
