% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_random.R
\name{gen.random.al}
\alias{gen.random.al}
\title{Generate random alteration landscapes}
\usage{
gen.random.al(al, N = 1000, params = NULL, r.seed = 100, ...)
}
\arguments{
\item{al}{The original alteration landscape}

\item{params}{parameter list}
}
\value{
A collection of random Alteration Landscapes
}
\description{
Generate random alteration landscapes
}
