% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_SpatialModel_projection_2010_package.R
\name{general_trends_2010}
\alias{general_trends_2010}
\title{Create a data frame of general trends for 2010}
\usage{
general_trends_2010(
  trainingGrids,
  general_trend_model_file = "Model_GeneralTrend.RData"
)
}
\arguments{
\item{trainingGrids}{data.frame; First nine fields in training data.}

\item{general_trend_model_file}{list; lm model of general trends}
}
\value{
estDF
}
\description{
Extract values from the general trends model and generate predictions
based off of the training dataset to create the estimates dataframe used
in update_pca_2010 function
}
