% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/script_SpatialModel_projection_2020&on_package.R
\name{general_trends_2020}
\alias{general_trends_2020}
\title{Create a data frame of general trends}
\usage{
general_trends_2020(
  updateGrids,
  general_trend_model_file = "Model_GeneralTrend.RData"
)
}
\arguments{
\item{updateGrids}{dataframe; updated inital dataset}

\item{general_trend_model_file}{list; lm model of general trends}
}
\value{
estDF
}
\description{
Extract values from the general trends model and generate predictions
based off of the training dataset to create the estimates dataframe used
in update_pca_2010 function
}
