% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL.R
\name{get.blocks}
\alias{get.blocks}
\title{Get sample/alteration blocks}
\usage{
get.blocks(al)
}
\arguments{
\item{al}{The alteration landscape}
}
\value{
Classification of samples and alterations in blocks.
}
\description{
Get sample/alteration blocks
}
