% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sl_stats.R
\name{get_sl_stats}
\alias{get_sl_stats}
\title{Get SuperLearner stats}
\usage{
get_sl_stats(sl_model, true_values, show_time = TRUE,
  show_weights = TRUE, plot_roc = TRUE)
}
\arguments{
\item{sl_model}{a \code{\link[SuperLearner]{SuperLearner}}}

\item{true_values}{(int) vector of correct outcome}

\item{show_time}{(lgl, default = TRUE) do you want training time will
be printed?}

\item{show_weights}{(lgl, default = TRUE) do you want waek lerner
risks and weigths will be printed?}

\item{plot_roc}{(lgl, default = TRUE), do you want the ROC plot will
be shown?}
}
\value{
(named double) AUC value (invisibly)
}
\description{
Get SuperLearner stats
}
\examples{
\dontrun{
    library(select)
    data("sais_selected", "SL.libraryD")
    SL.libraryD

    regressors <- as.data.frame(sais_selected[-12])
    outcome    <- as.integer(sais_selected[[12]] == "HbA1c <= -0.5")

    model <- train_sl(regressors, outcome)

    get_sl_stats(model, outcome)
}
}
