% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC_threshold.R
\name{get_thresh.2}
\alias{get_thresh.2}
\title{Establish which threshold is the best, using the 95\% of FP exclusion criterion}
\usage{
get_thresh.2(x, A)
}
\description{
Establish which threshold is the best, using the 95\% of FP exclusion criterion
}
