% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL.R
\name{get_vetos}
\alias{get_vetos}
\title{Build naive veto list by vetoing interactions of CNAs within the same chromosome}
\usage{
get_vetos(pool)
}
\arguments{
\item{pool}{list of CNAs in format AMP.chr..... or DEL.chr.....}
}
\value{
list of pairwise vetos
}
\description{
Build naive veto list by vetoing interactions of CNAs within the same chromosome
}
