% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_table.R
\name{interaction.table}
\alias{interaction.table}
\title{Assemble a table of pairwise interactions between alterations. Given N alterations, the table will contain N(N-1)/2 rows.}
\usage{
interaction.table(al, als)
}
\arguments{
\item{al}{Alteration Landscape}

\item{als}{AL statistics}
}
\value{
a dataframe
}
\description{
Assemble a table of pairwise interactions between alterations. Given N alterations, the table will contain N(N-1)/2 rows.
}
