% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{lis2mat}
\alias{lis2mat}
\title{Transform a pairwise table column into tabular form, taking care of NAs}
\usage{
lis2mat(lis, features = NULL, score = "score")
}
\arguments{
\item{lis}{table of pairwise interactions}

\item{features}{features to keep in the matrix}

\item{score}{name of the column to consider}
}
\value{
a 2d table
}
\description{
Transform a pairwise table column into tabular form, taking care of NAs
}
