% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{listmat_to_elemvec}
\alias{listmat_to_elemvec}
\title{Convert a list of m matrices (n x n) to n lists of n vectors of m elements}
\usage{
listmat_to_elemvec(temp)
}
\arguments{
\item{temp}{list of m matrices (n x n)}
}
\value{
n lists of n vectors of m elements
}
\description{
Convert a list of m matrices (n x n) to n lists of n vectors of m elements
}
