% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_FUNCTIONS.R
\name{LOGLIKS}
\alias{LOGLIKS}
\title{Evaluate log-likelihoods from full, null and fixed models}
\usage{
LOGLIKS(data, var.names, q.names = NULL, sumHauls = TRUE, fixed = NULL)
}
\description{
Evaluate log-likelihoods from full, null and fixed models
}
