% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlFuncs.R
\name{lselect}
\alias{lselect}
\title{Logistic selection curve}
\usage{
lselect(x, lenclass)
}
\arguments{
\item{x}{Vector containing parameters a (intercept) and b (shape)}

\item{lenclass}{Vector of lengths}
}
\value{
Vector of retention probabilities
}
\description{
The logistic selection curve function
}
