% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luad_data}
\alias{luad_data}
\title{Lung adenocarcinoma data from TCGA cohort}
\format{
A list object with GAM,sample & alteration annotations.
}
\usage{
data(luad_data)
}
\description{
A dataset containing the genome alteration matrix along with sample, alteration annotation as list object
}
\keyword{datasets}
