% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{mchoose}
\alias{mchoose}
\title{Calculate constant term in log-likelihood (including for non-integer counts)}
\usage{
mchoose(O, log = TRUE)
}
\arguments{
\item{O}{Matrix of observed counts}
}
\description{
Returns the constant term for binomial or multinomial counts.
The default is to return the logged constant.
The counts do not need to be integer valued.
}
