% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL.R
\name{new.AL}
\alias{new.AL}
\title{Init AL object}
\usage{
new.AL(am, samples.on.row = TRUE)
}
\arguments{
\item{am}{= NULL The binary alteration matrix}

\item{samples.on.row}{= TRUE Are samples on rows?}
}
\value{
An Alteration Landscape (AL) object with the gam
}
\description{
Create an Alteration Landscape (AL) object
}
\examples{
am = matrix(rbinom(50, 1, 0.1), ncol=5, nrow=10)
new.AL(am)

}
