% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{new.ALS}
\alias{new.ALS}
\title{Create an Alteration Landscape Statistics (ALS) object}
\usage{
new.ALS(al)
}
\arguments{
\item{al}{Alteration landscape}
}
\value{
An Alteration Landscape Statistics (ALS) object
}
\description{
Create an Alteration Landscape Statistics (ALS) object
}
