% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{new.AMS}
\alias{new.AMS}
\title{Create an Alteration Matrix Statistics (AMS) object}
\usage{
new.AMS(am)
}
\arguments{
\item{al}{Alteration matrix}
}
\value{
An Matrix Landscape Statistics (AMS) object
}
\description{
Create an Alteration Matrix Statistics (AMS) object
}
