% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELECT.R
\name{nllhood}
\alias{nllhood}
\title{Negative log-likelihood of the SELECT model (excluding constants)}
\usage{
nllhood(theta, Data, Meshsize, r, rel.power, penalty.func)
}
\description{
The general negative log-likelihood function. Provided for
completeness, but not intended for user use.
}
