% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{pairwise.p_MI}
\alias{pairwise.p_MI}
\title{Compute P vlaue for custom score}
\usage{
pairwise.p_MI(observed, random)
}
\arguments{
\item{observed}{Random Alteration landscape}

\item{random}{Random Alteration stats}
}
\value{
Custom Pairwise score for random ams
}
\description{
Compute P vlaue for custom score
}
