% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{permSELECT}
\alias{permSELECT}
\title{Permute catch data}
\usage{
permSELECT(
  data,
  var.names,
  statistic,
  haul = "haul",
  paired = NULL,
  nsim = 2,
  gear = NULL,
  block = NULL,
  ...
)
}
\description{
Permutes data in SELECT format and returns the value of the user
supplied function\code{statistic}in a nsim by length(statistic) matrix
}
