% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{permute.SELECT}
\alias{permute.SELECT}
\title{Implement permutation test}
\usage{
permute.SELECT(
  Data,
  statistic,
  N,
  SetID = "Haul",
  Freqs = c("nfine", "nwide"),
  ...
)
}
\description{
Permutation test on data in SELECT format
}
