% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Pope}
\alias{Pope}
\title{Alternate hauls haddock data}
\format{
Dataframe of dimension 24 by 3. Lengthclass (cm) in the first column,
and haddock frequencies in the codend and cover in columns two and three, respectively.
}
\usage{
data(Pope)
}
\description{
Alternate hauls haddock data used in Wileman et al. (1996) and Pope et al. (1975).
Codend mesh is 87 mm diamond.
}
\examples{
data(Pope)
}
\references{
Wileman et al. (1995) Manual of Methods of Measuring the Selectivity of Towed Fishing Gears.
ICES Cooerative Research Report, No 215.
}
\keyword{datasets}
