% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prologue}
\alias{prologue}
\title{PROLOGUE RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 473 rows and 263
variables.}
\source{
\url{
    https://datadryad.org/resource/doi:10.5061/dryad.qt743/2
}
}
\usage{
prologue
}
\description{
A dataset containing the data for the PROLOGUE Randomized Clinical
Trial (\url{https://www.ncbi.nlm.nih.gov/pubmed/27351380}).
}
\details{
The variabels \code{hypertension_adj}, \code{dislipidemia_adj}, and
\code{total_colesterol} were computed to be selected in the
same way as in the SAIS (adjusted) dataset. I.e.
dislipidemia is defined according to the
American Heart Associations criteria, while hypertension
is calculated accordingly to the 2017 American College of
Cardiology/American Heart Association Hypertension Guideline
definition. All this values were computed only at the
baseline.
}
\note{
Patient in the dataset are $463$ but there were one row per
patient per adverse event they has experienced.
The only differences in the rows for a same patient are in
the variables starting with "ae_" indicating information about
adverse events (ae).
If you do not need those information you can reduce the dataset
without loosing any information different from ae with the code
\code{distinct(select(prologue, -starts_with("ae_")))}.
}
\keyword{datasets}
