% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prologue_miced}
\alias{prologue_miced}
\title{PROLOGUE RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 473 rows and 228
variables.}
\source{
\url{
    https://datadryad.org/resource/doi:10.5061/dryad.qt743/2
}
}
\usage{
prologue_miced
}
\description{
Imputed version of the \code{\link{prologue}} database.
}
\details{
Variables representing dates and variables with more than
400 missing data were removed before the imputation proces.
}
\keyword{datasets}
