% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prologue_miced_selected}
\alias{prologue_miced_selected}
\title{Selected information from PROLOGUE RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 385 rows and 12
variables. (see details)}
\source{
\url{
    https://datadryad.org/resource/doi:10.5061/dryad.qt743/2
}
}
\usage{
prologue_miced_selected
}
\description{
A dataset containing the imputed (for the missing entries)
and selected baseline information for patient of the
\code{\link{prologue}} study (both arms). Only information of
patients analyzed in the original study (see source) were included.
}
\details{
The variabels selected are: \code{age}, \code{gender}, \code{bmi},
\code{hypertension_adj}, \code{dyslipidemia_adj}, \code{adiponectin},
\code{sbp}, \code{dbp}, \code{hba1c}, \code{fpg}, \code{ldl}, and \code{delta_hba1c}
(the outcome).\preformatted{     Imputation were performed using MICE algorithm with 20
     iterations. Only the overmentioned variables (with the
     exclusion of the outcome) were imputed. For the imputation,
    , All, the information in the full \code{\link{prologue}}
     dataset where used.
}
}
\keyword{datasets}
