% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelnCurveDefinitions.R
\name{propncurves}
\alias{propncurves}
\title{Relative retention probabilities.}
\usage{
propncurves(rtype)
}
\description{
Returns the function giving the retention probabilities in each of the gears.
This takes in to account the selection curve of each gear and the gear configuration
(covered-codend, trouser trawl, gillnet).
The user can edit to add new definitions.
}
\details{
These functions are used to calculate the matrix of relative retention probabilities.
E.g.,.
r=propncurves(rtype)
rmatrix=outer(lens,Meshsize,r,theta).
The use of the outer function means that when r is called,
both lens and Meshsize will be vectors of length equal to
the number of lengthclasses times number of meshes.
}
