% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC.R
\name{prune_invariants}
\alias{prune_invariants}
\title{Remove interactions with observed overlap strictly similar to expected overlap}
\usage{
prune_invariants(results)
}
\arguments{
\item{results}{table of pairwise interactions}
}
\value{
a dataframe without invariant interactions
}
\description{
Remove interactions with observed overlap strictly similar to expected overlap
}
