% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_table.R
\name{prune_vetoed}
\alias{prune_vetoed}
\title{Remove vetoed interactions (eg CNA between the same chromosome)}
\usage{
prune_vetoed(results, veto = NULL)
}
\arguments{
\item{results}{table of pairwise interactions}
}
\value{
a dataframe without vetoed interactions
}
\description{
Remove vetoed interactions (eg CNA between the same chromosome)
}
