% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{qAIC}
\alias{qAIC}
\title{Calculate quasi-AIC}
\usage{
qAIC(fit, OD, correct = F)
}
\description{
Adjusts the AIC of a fit for the given overdisperion \code{OD}.
AICc is used if \code{correct=T}.
}
