% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_FUNCTIONS.R
\name{QUARTIC}
\alias{QUARTIC}
\title{Fit model averaged 4th order polynomial}
\usage{
QUARTIC(
  Catch,
  vars = c("lgth", "nC", "nT"),
  Quasi = F,
  wgt = "AICc",
  All = TRUE
)
}
\arguments{
\item{Catch}{Matrix with data in first three columns}

\item{vars}{Column variable names if not vars=c("lgth","nC","nT")}

\item{Quasi}{Logical, whether to apply quasibinomial correction}

\item{wgt}{Weighting criterion. If not "AICc" then AIC is used}

\item{All}{If not TRUE then only full models from null to 4th order are used.}

\item{plens}{If specified, lengths at which to calculated fitted values.}
}
\value{
List containing fitted values, overdispersion, wgt and dredge averaged fit.
}
\description{
Fit model averaged 4th order polynomial
}
