% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MI.R
\name{r.al.pairwise.alteration.custom_measure}
\alias{r.al.pairwise.alteration.custom_measure}
\title{Compute custom score for the random ams}
\usage{
r.al.pairwise.alteration.custom_measure(
  als,
  r.als,
  n.cores = 1,
  verbose = TRUE,
  folder = "./",
  ...
)
}
\arguments{
\item{r.als}{Random Alteration stats}

\item{r.al}{Random Alteration landscape}

\item{cache}{constant data. Speed up calculations!}
}
\value{
Custom Pairwise score for random ams
}
\description{
Compute custom score for the random ams
}
