% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stats.R
\name{r.al.pairwise.alteration.stats}
\alias{r.al.pairwise.alteration.stats}
\title{Compute overlap statistics for the random ams}
\usage{
r.al.pairwise.alteration.stats(
  al,
  do.blocks = FALSE,
  n.cores = 1,
  verbose = TRUE,
  folder = "./"
)
}
\arguments{
\item{al}{Alteration landscape}
}
\value{
Pairwise statistics for random ams
}
\description{
Compute overlap statistics for the random ams
}
