% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{Randomize}
\alias{Randomize}
\title{Return a permuted SELECT format dataframe}
\usage{
Randomize(catch, varnames = c("Haul", "nwide", "nfine"))
}
\arguments{
\item{catch}{Stacked matrix or dataframe of catches in SELECT format}

\item{varnames}{Length-3 vector containing name of haulID variable, and 2 catch freq vars}
}
\value{
Dataframe, with randomized gear treatment within each haul
}
\description{
Randomization of gear type within each haul
}
