% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{RandomizeLong}
\alias{RandomizeLong}
\title{Return a dataframe with permutations of haul treatment}
\usage{
RandomizeLong(Data, haul = "Haul", Gear = "Gear", BlockID = NULL)
}
\arguments{
\item{Data}{Stacked matrix or dataframe of catches in long format. Data are not
assumed paired, so could be different length range within each haul.}

\item{haul}{Name of grouping variable containing the tow/haul/set id number.}

\item{Gear}{Name of the gear (treatment) indicator variable.}

\item{BlockID}{Permutation is within blocks}
}
\value{
Dataframe, with randomized gear treatment.
}
\description{
Randomization of haul treatments (may be more than 2)
}
