% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{RandomizeLong231105}
\alias{RandomizeLong231105}
\title{Return a dataframe with permutations of haul treatment}
\usage{
RandomizeLong231105(Data, SetID = "Haul", Gear = "Gear")
}
\arguments{
\item{Data}{Stacked matrix or dataframe of catches in long format. Data are not
assumed paired, so could be different length range within each haul.}

\item{SetID}{Name of grouping variable containing the tow/haul/set id number.}

\item{GearID}{Name of the gear (treatment) indicator variable.}
}
\value{
Dataframe, with randomized gear treatment.
}
\description{
Randomization of haul treatments (may be more than 2)
}
