% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{raster_slope}
\alias{raster_slope}
\title{calculate slope}
\usage{
raster_slope(attribute_tble, variable_arr, centroid_shp)
}
\value{
output_slope
}
\description{
calculate slope
}
\details{
create slope dataframe from rasters
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
