% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{Rep.OD}
\alias{Rep.OD}
\title{Calculate the replication estimate of overdispersion, REP.}
\usage{
Rep.OD(data, var.names, q.names = NULL, minE = 1)
}
\arguments{
\item{data}{Catch data in SELECT format}

\item{var.names}{Character vector giving the variable names.
The first name is always the length variable,
followed by names of the raw catch variables. E.g., c("lgth","nA","nB")}

\item{minE}{Numeric value giving the minimum fitted count for an observation
to be used in the Pearson chi-square.}

\item{q.name}{Character vector giving the names of sampling fractions, if any.}
}
\value{
dataframe
}
\description{
REP is a simple estimation of overdispersion calculated
from stacked combined-hauls data based on fitting.
It should not be used on unpaired data.
a full model (one term for each length) to stacked combined-hauls data.
It it based on the Pearson chi-square estimated of overdisperion
from fitting the full model ((one term for each length).
For robustness against sparsity, it is
important that only counts with a minimum specified fitted value are used.
See Millar and Fryer (1999, Estimating the size-selection curves of
towed gears, traps, nets and hooks) for details.
}
