% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlFuncs.R
\name{Rep.ttfit}
\alias{Rep.ttfit}
\title{Replicate estimation of over-dispersion for paired hauls data.}
\usage{
Rep.ttfit(
  catch = catchdat,
  type = "logit",
  x0 = c(-9, 0.3, 0.5),
  delta = 1,
  ind.psplit = T,
  suff.big = 3,
  plots = F,
  details = T,
  verbose = T
)
}
\arguments{
\item{catch}{Matrix with 4 columns. These must be
\enumerate{
\item length (i.e., midpoint of length class);
\item Freq in experimental gear; 3) Freq in nonselective control gear
\item Haul ID
}}

\item{type}{Character. Use type="rich" to fit Richards curve}

\item{details}{If TRUE, returns additional output}

\item{verbose}{If FALSE then messages are suppressed}
}
\description{
Replicate estimation of over-dispersion for logistic or Richards
selection curve fitted to stacked (by individual haul) paired hauls data.
Pairing would typically be via twin hauls, alternate hauls or trouser trawls.
This function applies the REP correction of over-dispersion in
Millar et al. (2004, Modelling between-haul variability in the size selectivity of trawls,
Fisheries Research. 67: 171-181.)
}
