% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrawlFuncs.R
\name{rselect}
\alias{rselect}
\title{Richards selection curve}
\usage{
rselect(x, lenclass)
}
\arguments{
\item{x}{Vector containing parameters a (intercept), b (shape) and delta}

\item{lenclass}{Vector of lengths}
}
\value{
Vector of retention probabilities
}
\description{
The Richards selection curve function
}
