% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{SafeSample}
\alias{SafeSample}
\title{Avoid issue with sample from a single value}
\usage{
SafeSample(x, size = length(x), replace = T)
}
\description{
Avoid issue with sample from a single value
}
