% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sais}
\alias{sais}
\title{SAIS RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 208 rows and 46
variables. (see details)}
\source{
\url{https://doi.org/10.1371/journal.pone.0164255.s004} and
\url{https://doi.org/10.1371/journal.pone.0164255.s005}
}
\usage{
sais
}
\description{
A dataset containing the data for the SAIS Randomized Clinical
Trial (\url{https://doi.org/10.1371/journal.pone.0164255.s004} for
the baseline data, and
\url{https://doi.org/10.1371/journal.pone.0164255.s005} for the
follow-up data.
}
\details{
The variabels \code{hypertension_adj}, \code{dislipidemia_adj}, and
\code{total_colesterol} were computed to be selected in the
same way as in the PROLOGUE (adjusted) dataset. I.e.
dislipidemia is defined according to the
American Heart Associations criteria, while hypertension
is calculated accordingly to the 2017 American College of
Cardiology/American Heart Association Hypertension Guideline
definition. All this values were computed only at the
baseline.
}
\note{
Patient in the dataset are $104$ but there were one row per
time (i.e., baseline or follow-up (AKA "fup")).
}
\keyword{datasets}
