% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sais_selected}
\alias{sais_selected}
\title{Selected information from SAIS RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 43 rows and 15
variables. (see details)}
\source{
\url{https://doi.org/10.1371/journal.pone.0164255.s004} and
\url{https://doi.org/10.1371/journal.pone.0164255.s005}
}
\usage{
sais_selected
}
\description{
A dataset containing the selected baseline information for patient of
the \code{\link{sais}} study in the sitagliptin arm.
}
\details{
The variabels selected are: \code{age}, \code{gender}, \code{bmi},
\code{hypertension_adj}, \code{dyslipidemia_adj}, \code{adiponectin},
\code{sbp}, \code{dbp}, \code{hba1c}, \code{fpg}, \code{ldl}, and \code{delta_hba1c}
(the outcome)
}
\keyword{datasets}
