% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-package.R
\docType{package}
\name{select-package}
\alias{select}
\alias{select-package}
\title{select: SuperLearner for randomized Clinical Trial support}
\description{
Provide the code for the application of a
    SuperLearner approach in support to Randomized Clinical Trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/select}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/select/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

Authors:
\itemize{
  \item Ileana Baldi \email{ileana.baldi@unipd.it}
  \item Paola Berchialla \email{paola.berchialla@unito.it}
  \item Dario Gregori \email{dario.gregori@unipd.it}
}

}
\keyword{internal}
