% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{SELECT.FORMAT}
\alias{SELECT.FORMAT}
\title{Change long format to SELECT format}
\usage{
SELECT.FORMAT(
  Df,
  by = c("haul", "lgth"),
  gear = "gear",
  freq = "freq",
  q.name = NULL,
  paired = T
)
}
\arguments{
\item{by}{Character vector with names of the variables to join by,
typically TowID and lgth}

\item{gear}{Character giving gear variable name}

\item{freq}{Character giving frequency variable name}

\item{q.name}{Name of sampling fraction variable}

\item{paired}{Are data paired within gear?}
}
\value{
Dataframe, in SELECT format
}
\description{
Change long format to SELECT format
}
