% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.internal}
\alias{select.internal}
\title{internal SELECT routine}
\usage{
select.internal(
  al,
  r.al,
  folder = "./",
  n.cores = 40,
  vetos = NULL,
  rho = 0.1,
  lambda = 15,
  save.intermediate.files = TRUE,
  calculate_APC_threshold = TRUE,
  calculate_FDR = TRUE,
  verbose = TRUE,
  nbatches = 1,
  FDR.thresh = 0.1
)
}
\value{
selection interactions between alterations

#@examples
#select()
}
\description{
Run SELECT analysis on input al. All params and null model must be set and valid
}
