% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.on.AL.R
\name{select.on.AL}
\alias{select.on.AL}
\title{SELECT analysis on an AL with null model}
\usage{
select.on.AL(
  al,
  folder = "./",
  r.seed = 110,
  n.cores = 40,
  vetos = NULL,
  min.feature.support = 5,
  min.feature.freq = 0.001,
  remove.0.samples = TRUE,
  remove.unknown.class.samples = TRUE,
  rho = 0.1,
  lambda = 15,
  save.intermediate.files = TRUE,
  calculate_APC_threshold = TRUE,
  calculate_FDR = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{al}{Input AL with precalculated null model}
}
\value{
selection interactions between alterations

#@examples
#select.on.AL()
}
\description{
Run SELECT analysis on input AL
}
