% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELECT.R
\name{SELECT}
\alias{SELECT}
\title{Fit selection curves to fishing gears using the SELECT method.}
\usage{
SELECT(
  Data,
  dtype,
  stype = "logistic",
  Q = NULL,
  Meshsize = NULL,
  x0 = NULL,
  rel.power = NULL,
  penalty.func = NULL,
  print.out = T,
  control = list(maxit = 10000, reltol = 1e-08),
  Fit = T
)
}
\description{
SELECT is an acronym for the Share Each LEngths Catch Total method
of Millar (1992). SELECT is a general statistical framework to  estimate the
selectivity of trawls, hooks and gillnets from experimental data.
}
\references{
Millar, R. B. (1992).
Estimating the size-selectivity of fishing gear by conditioning on the
total catch. Journal of the American Statistical Association. 87: 962-968.
}
