% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{SELECT_FORMAT}
\alias{SELECT_FORMAT}
\title{Put length-freqs stacked by gear (and/or other variables) into SELECT format.
Default assumes input df has columns named "TowID", "lgth", "freq" and "gear"
where gear may be multilevel.}
\usage{
SELECT_FORMAT(Df, by = c("TowID", "lgth"), gear = "gear", freq = "freq")
}
\arguments{
\item{by}{Character vector with names of the variables to join by,
typically TowID and lgth}

\item{gear}{Character giving gear variable name}

\item{freq}{Character giving frequency variable name}
}
\value{
Dataframe, in SELECT format
}
\description{
Change long format to SELECT format
}
