% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CURVE_DEFNS.R
\name{selncurves}
\alias{selncurves}
\title{Retention curve definitions.}
\usage{
selncurves(rtype)
}
\description{
The definitions of the retention/selection curves supported by SELECT. The user
can edit to add new definitions.
}
\details{
These curves are used by function propncurves, to calculate the matrix of relative
retention probabilities as generated by the R code
rmatrix=outer(lens,Meshsize,r,theta).
This means that when r is called,
both lens and Meshsize will be vectors of length equal to
the number of lengthclasses times the number of meshes.
}
