% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serpentine-data.R
\docType{data}
\name{serpentine}
\alias{serpentine}
\title{Functional trait data for species in serpentine grassland}
\format{dataframe of 48 species and 2 traits}
\source{
Funk, J.L. & Wolf, A.A. 2016. Testing the trait based community framework: Do functional traits predict competitive outcomes? Ecology 97:2206-2211.
}
\usage{
data(serpentine)
}
\description{
A dataset containing two traits for species in a serpentine grassland.
The variables are as follows:
}
\details{
\itemize{
  \item species. species name
  \item wue. average water use efficiency, rate of carbon assimilation per unit of water used
  \item rootdepth. average rooting depth in soil
}
}
\keyword{datasets}
