% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{SimCatch}
\alias{SimCatch}
\title{Permuation example data}
\format{
dataframe \code{SimCatch} with 1000 rows and 4 columns.
}
\usage{
data(SimCatch)
}
\description{
Simulated catch data from gears A and B. Used in Appendix 2 of Millar (2024) to
demonstrate use of permutation testing.
}
\examples{
data(SimCatch)
}
\references{
Millar R. B. (2024). Incorrect inference from size-selectivity studies
due to widespread misuse of bootstrap intervals.
}
\keyword{datasets}
