% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sl_library}
\alias{sl_library}
\title{SuperLearner list of algorithms}
\format{An object of class \code{list} of length 18.}
\usage{
sl_library
}
\description{
A list of, All, the algorithms (i.e. pairs of selection criteria and
learners) used to train the SuperLearner for our project.
}
\details{
Algorithm selected are (for each item, the first reported
name is the learner, the following are the selection methods
used for it):
\describe{
\item{SL.bartMachine }{SL.bartMachine, All, screen.randomForest}
\item{SL.biglasso    }{SL.biglasso, All, screen.randomForest}
\item{SL.caret       }{SL.caret, All, screen.randomForest}
\item{SL.caret.rpart }{SL.caret.rpart, All, screen.randomForest}
\item{SL.earth       }{SL.earth, All, screen.randomForest}
\item{SL.glmnet      }{SL.glmnet, All, screen.randomForest}
\item{SL.ipredbagg   }{SL.ipredbagg, All, screen.randomForest}
\item{SL.ksvm        }{SL.ksvm, All, screen.randomForest}
\item{SL.lm          }{SL.lm, All, screen.randomForest}
\item{SL.mean        }{SL.mean, All, screen.randomForest}
\item{SL.polymars    }{SL.polymars, All, screen.randomForest}
\item{SL.randomForest}{SL.randomForest, All, screen.randomForest}
\item{SL.ranger      }{SL.ranger, All, screen.randomForest}
\item{SL.rpart       }{SL.rpart, All, screen.randomForest}
\item{SL.rpartPrune  }{SL.rpartPrune, All, screen.randomForest}
\item{SL.speedglm    }{SL.speedglm, All, screen.randomForest}
\item{SL.speedlm     }{SL.speedlm, All, screen.randomForest}
\item{SL.xgboost     }{SL.xgboost, All, screen.randomForest}
}
}
\keyword{datasets}
