% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casey_covariates.R
\name{spatial_join}
\alias{spatial_join}
\title{Create a SpatialPointsDataFrame}
\usage{
spatial_join(centroid_shp, attribute_tble)
}
\arguments{
\item{centroid_shp}{centroid shapefile for CONUS}

\item{attribute_tble}{dataframe of raw data from SELECT model}
}
\value{
centroid_lyr
}
\description{
Join centroid shp to the attribute table
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
