% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_FUNCTIONS.R
\name{SplineSELECT}
\alias{SplineSELECT}
\title{Fit smooth function to catch share data}
\usage{
SplineSELECT(
  data,
  var.names,
  q.names = NULL,
  useTots = TRUE,
  bs = "cr",
  k = 5,
  sp = NULL,
  quasi = FALSE,
  rm.zeros = TRUE
)
}
\arguments{
\item{bs}{Choice of smoother}

\item{k}{Dimension of the basis. k=3 in minimum for natural cubic spline.}

\item{sp}{Supplied smoothing parameter}

\item{quasi}{Logical, whether to apply quasibinomial correction}

\item{Catch}{Matrix with data in first three columns}

\item{m}{Order of the penalty}
}
\value{
List containing fitted model.
}
\description{
Fit spline to catch share - description needs updating
}
