% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{SplineStatistics}
\alias{SplineStatistics}
\title{Compute permutation test statistics}
\usage{
SplineStatistics(SplineFit, MLS = NULL)
}
\description{
\code{SplineStatistics} calculates a variety of test statistics for use
with permutation testing. The appropriate test statistic(s) to use will depend on
the questions of interest. It does NOT assume that the SplineSELECT fit
returns a log-likelihood and so still work when a quasi=T fit is used.
}
