% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{StartValues}
\alias{StartValues}
\title{Initial parameter values for logistic and Richards curves}
\usage{
StartValues(rtype, Data)
}
\description{
Returns crude data-driven starting values for logistic and
Richards curves
}
