% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{T45SidePanels}
\alias{T45SidePanels}
\title{T45 side panel data}
\format{
dataframe \code{T45SidePanels} with 252 rows and 8 columns.
}
\usage{
data(T45SidePanels)
}
\description{
School prawn catch data (including sampling fractions) from twin trawls with
panaeid gears modified with either
a 32 mm or 35 mm T45 (square orientation of mesh) side panels.
Used in Appendix 2 of Millar (2024) to demonstrate use of permutation testing.
}
\examples{
data(T45SidePanels)
}
\references{
Broadhurst et al. (2018). T45 side panels improve penaeid-trawl selection.
}
\keyword{datasets}
