% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC_threshold.R
\name{TP.FP.stat}
\alias{TP.FP.stat}
\title{Derive TP and FP stats for APC based on FDR scores for a specific threshold}
\usage{
TP.FP.stat(A, th)
}
\description{
Derive TP and FP stats for APC based on FDR scores for a specific threshold
}
