% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_APC_threshold.R
\name{TP.FP.stats}
\alias{TP.FP.stats}
\title{Derive TP and FP stats for APC based on FDR scores for a range of thresholds}
\usage{
TP.FP.stats(A, res = 1000)
}
\description{
Derive TP and FP stats for APC based on FDR scores for a range of thresholds
}
