% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sl.R
\name{train_sl}
\alias{train_sl}
\title{wrap for training a SuperLearner}
\usage{
train_sl(training_data, true_values, SL.libraryD = SL.libraryD,
  method = "method.NNLS", seed = 123)
}
\arguments{
\item{training_data}{(data.frame) predictors data}

\item{true_values}{(int) vector of outcome}

\item{SL.libraryD}{Either a character vector of prediction algorithms
a list containing character vectors.}

\item{method}{(default = "method.NNLS") A list (or a function to
create a list) containing details on estimating the
coefficients for the super learner and the model to
combine the individual algorithms in the library.}

\item{seed}{(default = 123) the seed for reproducibility}
}
\value{
an object of class \code{\link[SuperLearner]{SuperLearner}}
}
\description{
wrap for training a SuperLearner
}
\examples{
\dontrun{
    library(select)
    data("sais_selected", "SL.libraryD")
    SL.libraryD

    fit_sl(
        as.data.frame(sais_selected[-12]),
        as.integer(sais_selected[[12]] == "HbA1c <= -0.5")
    )
}
}
