% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropicalforest-data.R
\docType{data}
\name{tropicalforest}
\alias{tropicalforest}
\title{Functional trait data for tree species in tropical forest}
\format{dataframe of 41 species and 2 traits}
\source{
McCarthy, James K. 2018. Predicting the diversity and functional composition of woody plant communities under climate change PhD Thesis, School of Biological Sciences, The University of Queensland. Advisor: Dr. John Dwyer. doi:10.14264/uql.2018.134
}
\usage{
data(tropicalforest)
}
\description{
A dataset containing two traits for species in a tropical forest.
The variables are as follows:
}
\details{
\itemize{
  \item species. species name
  \item sla. specific leaf area
  \item wd. wood density
}
}
\keyword{datasets}
