% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Trout}
\alias{Trout}
\title{Gillnet data}
\format{
list object with components Counts (18 by 6 - first column is lengths) and
Meshsize (vector of meshsizes)  respectively.
}
\usage{
data(Trout)
}
\description{
Counts of trout from a five mesh gang of gillnets,
from Table 1 of Hansen et al. (1997)
}
\examples{
data(Trout)
}
\references{
Hansen et al. (1997). Gillnet selectivity for lake trout
(Salvelinus namaycush) in Lake Superior.
Canadian Journal of Fisheries and Aquatic Sciences, 54: 2483-2490.
}
\keyword{datasets}
