% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_SpatialModel_projection_2010_package.R
\name{update_pca_2010}
\alias{update_pca_2010}
\title{Update data from PCA model predictions for 2010}
\usage{
update_pca_2010(
  trainingGrids,
  estDF,
  pcaModel = "Model_BuPca.RData",
  selectedFeatures = "temp_SelectedFeatures_projection.RData",
  maskGrids = "data_FinalMask.csv",
  popGrids2000 = "data_2000TotalPop_SSPBaseYr.csv"
)
}
\arguments{
\item{trainingGrids}{dataframe; inital dataframe}

\item{estDF}{dataframe; created by general_trends()}

\item{pcaModel}{prcomp object; PCA model of BU data}

\item{selectedFeatures}{dataframe; selected features of model outputs}

\item{maskGrids}{dataframe; gridded mask data}

\item{popGrids2000}{dataframe; gridded population data 2000}

\item{BUCovars}{dataframe; dataframe of focal statistics for 3 variables with moving window (3,5,7,9)}
}
\value{
list; extracted components from update Grids data => list(bu_covars, selectedFeatures, estDF)
}
\description{
Predict values for BUCovars dataset given the pcaModel and
use to update the BUCovars dataframe, estDF dataframe and create
the selectedFeatures dataframe
}
