% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/script_SpatialModel_projection_2020&on_package.R
\name{update_pca_2020}
\alias{update_pca_2020}
\title{Update data from PCA model predictions for 2020}
\usage{
update_pca_2020(
  updateGrids,
  estDF,
  BUCovars = "temp_tbl_attr_old.RData",
  pcaModel = "Model_BuPca.RData",
  selectedFeatures = "temp_SelectedFeatures_projection.RData"
)
}
\arguments{
\item{updateGrids}{dataframe; updated inital dataset}

\item{estDF}{dataframe; created by general_trends()}

\item{BUCovars}{dataframe; of focal statistics for 3 variables with moving window (3,5,7,9)}

\item{pcaModel}{prcomp object; PCA model of BU data}

\item{selectedFeatures}{dataframe; selected features of model outputs}
}
\value{
list; extracted components from update Grids data => list(bu_covars, selectedFeatures)
}
\description{
Predict values for BUCovars dataset given the pcaModel and
use to update the BUCovars dataframe, estDF dataframe and create
the selectedFeatures dataframe
}
