% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{WgtAvg}
\alias{WgtAvg}
\title{Crude weighted average (0.25,0.5,0.25) with immediate neighbours}
\usage{
WgtAvg(y, w = c(0.25, 0.5, 0.25))
}
\description{
Crude weighted average (0.25,0.5,0.25) with immediate neighbours
}
