% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{bootSELECT}
\alias{bootSELECT}
\title{Bootstrap catch data}
\usage{
bootSELECT(
  data,
  var.names,
  statistic,
  haul = NULL,
  nsim = 2,
  paired = NULL,
  block = NULL,
  gear = NULL,
  within.resamp = TRUE,
  verbose = 1,
  ...
)
}
\arguments{
\item{data}{Stacked matrix or dataframe of catches in SELECT format,}

\item{haul}{Name of grouping variable containing the tow/haul/set id number.
Vector of same length as number of rows in Data.}

\item{paired}{Logical True if the data are paired}

\item{block}{If specified, name of blocking variable.
Bootstrapping is first done over blocks, and then sets within each block.}

\item{gear}{If specified, name of the gear indicator variable.
This restricts bootstrapping of sets to within each gear and is
for use with non-paired data.}

\item{verbose}{If set to 0 will suppress messages and progress bar.
If >1 will also print the value of \code{statistic} for the observed \code{data}.}

\item{...}{Other parameters to be passed to the \code{statistic} function. E.g.,
q.names (sampling fractions) if applicable.}
}
\description{
Applies a double bootstrap to data in SELECT format and
evaluates the vector valued function \code{statistic}. The returned value is a
nsim by length(statistic) matrix of bootstrap statistics.
}
