% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{Dble.boot}
\alias{Dble.boot}
\title{Return a dataframe containing double bootstrapped raw data.}
\usage{
Dble.boot(Data, SetID = "Haul", Freqs = c("nfine", "nwide"), smooth = T)
}
\arguments{
\item{Data}{Stacked matrix or dataframe of catches in SELECT format,
with lengthclass in first column. Remaining columns are raw catch frequencies}

\item{smooth}{Smooth at within-haul phase to avoid losing degrees of freedom (from increasing number of zero freqs)}

\item{Set.id}{Grouping variable containing the tow/haul/set id number.
Vector of same length as number of rows in Data.}
}
\value{
Dataframe of double-bootstrapped freqs
}
\description{
Double bootstrap function used by boot.SELECT function
}
