% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{Dble.boot}
\alias{Dble.boot}
\title{Return a SELECT format dataframe containing hierarchical (double) bootstrapped catch data.}
\usage{
Dble.boot(
  data,
  haul = "Haul",
  block = NULL,
  gear = NULL,
  Freqs = c("nfine", "nwide"),
  paired = T,
  within.resamp = T,
  smooth = F
)
}
\arguments{
\item{data}{Stacked matrix or dataframe of catches in SELECT format,
with lengthclass in first column. Remaining columns are raw catch frequencies}

\item{haul}{Name of grouping variable containing the tow/haul/set id number.
Vector of same length as number of rows in Data.}

\item{block}{If specified, name of blocking variable.
Bootstrapping is first done over blocks, and then sets within each block.}

\item{gear}{If specified, name of the gear indicator variable.
This restricts bootstrapping of sets to within each gear and is intended to be used
for non-paired data.}

\item{paired}{Logical. True if the data are paired}

\item{within.resamp}{If F, no resampling at observation level.}

\item{smooth}{Smooth at within-haul phase to avoid losing degrees of freedom (from increasing number of zero freqs)}
}
\value{
A list object with two components. The first is the bootstrapped dataframe in SELECT format, and the second is a vector containing the bootstrapped \code{haul} IDs.
and
}
\description{
Double bootstrap function used by bootSELECT function
}
