% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{Dble.boot230425}
\alias{Dble.boot230425}
\title{Old workhorse}
\usage{
Dble.boot230425(
  Data,
  haul = "Haul",
  BlockID = NULL,
  Freqs = c("nfine", "nwide"),
  inner.resamp = T,
  smooth = F
)
}
\description{
Old workhorse
}
