% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapAndRandomize.R
\name{delta.method}
\alias{delta.method}
\title{Modification of deltamethod (msm) that also returns function value}
\usage{
delta.method(g, mean, cov, ses = TRUE)
}
\description{
Modification of deltamethod (msm) that also returns function value
}
