% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{Estimates}
\alias{Estimates}
\title{Returns selectivity parameters}
\usage{
Estimates(fit)
}
\description{
Calculates the estimates and standard error of selectivity
curve implemented by SELECT. The user can edit to add new calculations for
any user-supplied retention curves.
}
