% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{focal_stats}
\alias{focal_stats}
\title{create dataframe of focal statistica from variable arrangement in attribute table}
\usage{
focal_stats(attribute_tble, variable_arr, size)
}
\value{
focal_data
}
\description{
create dataframe of focal statistica from variable arrangement in attribute table
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
