% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{permSELECT}
\alias{permSELECT}
\title{Sample permutation distribution of a user-defined statistic}
\usage{
permSELECT(
  data,
  var.names,
  q.names = NULL,
  statistic,
  haul = "haul",
  paired = NULL,
  nsim = 2,
  block = NULL,
  gear = NULL,
  verbose = 1
)
}
\arguments{
\item{data}{Stacked matrix or dataframe of catches in SELECT format.}

\item{var.names}{Character vector of length 3 containing the names of the length variable and catch variables.
#' @param q.names Character vector of length 2 containing the names of the sampling fractions.
#' @param statistic The numeric or vector-valued function to be applied to the permuted data. This function would typically be a fit or test statistic.}

\item{haul}{Name of the grouping variable identifying the haul. This could be a paired-haul in the case of twin or alternate design, in which case both gears must share the same haul identifier.}

\item{paired}{Logical. This is a required parameter. Set to \code{TRUE} if the data are paired, \code{FALSE} otherwise.}

\item{nsim}{Number of permutations  to be performed.}

\item{block}{If specified, name of blocking variable. For example, day of deployment.
Permuting is then restricted to being within each block.}

\item{gear}{If specified, name of the gear indicator variable.
This is required for use with non-paired data.}
}
\value{
A matrix of dimension \code{nsim} by \code{length(statistic)} containing the bootstrap
}
\description{
\code{permSELECT} permutes catch data in SELECT format and returns the value of the user
supplied function\code{statistic}. The returned value is a \code{nsim} by \code{length(statistic)} matrix of statistics.

The argument \code{block} is used to restrict permuting to within blocks.
For example, the blocking variable could be day, or region. See details below.

The experiment may be paired of unpaired. In the latter case the \code{SELECT} format
dataframe must contain a variable that indicates the gear type and this variable is
used as the \verb{gear=} argument.
}
