% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP&PERMUTATION.R
\name{permSELECT}
\alias{permSELECT}
\title{Permute catch data}
\usage{
permSELECT(obj, statistic, haul = "Haul", nsim = 2, ...)
}
\description{
Permutes data in SELECT format and returns the value of the user
supplied function\code{statistic}in a nsim by length(statistic) matrix
}
