% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{PlotCurves}
\alias{PlotCurves}
\title{Plots the fitted curves and returns fitted values at specified lengths}
\usage{
PlotCurves(
  fit,
  plotlens = NULL,
  Meshsize = NULL,
  rel.power = NULL,
  standardize = F,
  plot.out = T,
  xlab = "Length (cm)",
  ylab = "Retention curve",
  ...
)
}
\description{
Plot the fitted selectivity curve(s)
}
