% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_FUNCTIONS.R
\name{PolySELECT}
\alias{PolySELECT}
\title{Fit model averaged 4th order polynomial}
\usage{
PolySELECT(
  data,
  var.names,
  q.names = NULL,
  sumHauls = TRUE,
  q.ODadjust = T,
  quasi = F,
  wgt = "AICc",
  All = TRUE
)
}
\arguments{
\item{data}{Matrix with data in first three columns}

\item{wgt}{Weighting criterion. If not "AICc" then AIC is used}

\item{All}{If not TRUE then only full models from null to 4th order are used.}

\item{vars}{Column variable names if not vars=c("lgth","nC","nT")}

\item{plens}{If specified, lengths at which to calculated fitted values.}

\item{Quasi}{Logical, whether to apply quasibinomial correction}
}
\value{
List containing fitted values, overdispersion, wgt and dredge averaged fit.
}
\description{
Fit model averaged 4th order polynomial
}
