% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casey_covariates.R
\name{raster_slope}
\alias{raster_slope}
\title{Create slope dataframe for variable arrangment}
\usage{
raster_slope(attribute_tble, variable_arr, centroid_shp)
}
\arguments{
\item{attribute_tble}{dataframe of raw data from select model}

\item{variable_arr}{vector of variables listed in attribute table}

\item{centroid_shp}{centroid shapefile for CONUS}
}
\value{
output_slope
}
\description{
Calculate slope using landsat slopeasp function from SELECT model
attribute table, then convert to raster and extract data at the centroids
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
