% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_FUNCTIONS.R
\name{Raw2Tots}
\alias{Raw2Tots}
\title{Convert raw counts by haul to totals, and adjusts for any subsampling}
\usage{
Raw2Tots(data, var.names, q.names = NULL, useTots = T)
}
\description{
Convert raw counts by haul to totals, and adjusts for any subsampling
}
