% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sais}
\alias{sais}
\title{SAIS RCT data}
\format{A \link[tibble:tibble-package]{tibble} with 208 rows and 43
variables.}
\source{
\url{https://doi.org/10.1371/journal.pone.0164255.s004} and
\url{https://doi.org/10.1371/journal.pone.0164255.s005}
}
\usage{
sais
}
\description{
A dataset containing the data for the SAIS Randomized Clinical
Trial (\url{https://doi.org/10.1371/journal.pone.0164255.s004} per
i dati baseline, e
\url{https://doi.org/10.1371/journal.pone.0164255.s005} per i dati di
follow-up).
}
\note{
Patient in the dataset are $104$ but there were one row per
time (i.e., baseline or follow-up (AKA "fup")).
}
\keyword{datasets}
