% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{SELECT analysis}
\usage{
select(
  M,
  sample.class,
  alteration.class,
  folder = "./",
  r.seed = 110,
  n.cores = 40,
  vetos = NULL,
  n.permut = 1000,
  min.feature.support = 5,
  min.feature.freq = 0.001,
  remove.0.samples = TRUE,
  remove.unknown.class.samples = TRUE,
  rho = 0.1,
  lambda = 15,
  save.intermediate.files = TRUE,
  randomization.switch.threshold = 30,
  max.memory.size = 100,
  calculate_APC_threshold = TRUE,
  calculate_FDR = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{M}{gam}
}
\value{
selection interactions between alterations

#@examples
#select()
}
\description{
Run SELECT analysis on input gam
}
