% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_FUNCTIONS.R
\name{SELECT}
\alias{SELECT}
\title{Fit selection curves to fishing gears using the SELECT method.}
\usage{
SELECT(
  data,
  var.names,
  dtype = "cc",
  stype = "logistic",
  useTots = TRUE,
  q.names = NULL,
  Meshsize = NULL,
  x0 = NULL,
  rel.power = NULL,
  penalty.func = NULL,
  verbose = NULL,
  control = list(maxit = 10000, reltol = 1e-08),
  Fit = TRUE
)
}
\arguments{
\item{data}{Catch data in SELECT format}

\item{var.names}{Character vector giving the variable names.
The first name is always the length variable,
followed by names of the raw catch variables. E.g., c("lgth","nA","nB")}

\item{q.name}{Character vector giving the names of sampling fractions, if any.}
}
\description{
SELECT is an acronym for the Share Each LEngths Catch Total method
of Millar (1992). SELECT is a general statistical framework to  estimate the
selectivity of trawls, hooks and gillnets from experimental data.
}
\references{
Millar, R. B. (1992).
Estimating the size-selectivity of fishing gear by conditioning on the
total catch. Journal of the American Statistical Association. 87: 962-968.
}
