% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{spatial_join}
\alias{spatial_join}
\title{join the grid center layer to attribute feature file}
\usage{
spatial_join(centroid_shp, attribute_tble)
}
\value{
centroid_lyr
}
\description{
join the grid center layer to attribute feature file
}
\author{
Casey R. McGrath (casey.mcgrath@pnnl.gov)
}
