% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Trout}
\alias{Trout}
\title{Gillnet data}
\format{
list object with components Counts (18 by 6 - first column is lengths) and
Meshsize (vector of meshsizes)  respectively.
}
\usage{
data(Trout)
}
\description{
Counts of trout from a five mesh gang of gillnets
}
\examples{
data(Trout)
}
\references{
Helser T. E., et al. (1963). Estimating gillnet selectivity
using nonlinear response surface regression.
Can. J. Fish. Aquat. Sci. 55: 1328-1337.
}
\keyword{datasets}
